# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Job Tracker API
  version: 0.1.0
  description: |
    Track job applications: vacancies, applications moving through a Kanban
    pipeline, and the account that owns them.

    Every route requires a session unless it is marked `security: []`. A
    session is sent either as `Authorization: Bearer <token>` or as the
    HttpOnly `session` cookie set by the login endpoints. A request without a
    valid session gets 401, never 403.

    Request bodies must be sent with `Content-Type: application/json`
    (otherwise 415), must not contain unknown fields (otherwise 400), and are
    capped at 1MB.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
  - sessionCookie: []
...
paths:
...
  /vacancies:
    get:
      tags: [vacancies]
      operationId: listVacancies
      summary: List vacancies
      parameters:
        - name: company_id
          in: query
          schema: { type: integer, format: int64 }
        - name: salary_from
          in: query
          schema: { type: integer, format: int32 }
        - name: has_application
          in: query
          description: Omit for any, `true` for vacancies with an active application, `false` for those without.
          schema: { type: boolean }
        - name: work_mode
          in: query
          description: Comma-separated.
          style: form
          explode: false
          schema:
            type: array
            items: { $ref: "#/components/schemas/WorkMode" }
        - name: employment_type
          in: query
          description: Comma-separated.
          style: form
          explode: false
          schema:
            type: array
            items: { $ref: "#/components/schemas/EmploymentType" }
        - name: language
          in: query
          description: Comma-separated.
          style: form
          explode: false
          schema:
            type: array
            items: { type: string }
        - name: source
          in: query
          description: Comma-separated.
          style: form
          explode: false
          schema:
            type: array
            items: { type: string }
        - name: q
          in: query
          description: Matches title or description.
          schema: { type: string }
        - name: sort
          in: query
          description: Defaults to `created_at`, newest first.
          schema:
            type: string
            enum: [created_at, updated_at, posted_at, title, salary]
        - $ref: "#/components/parameters/Desc"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: A page of vacancies.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VacancyList"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [vacancies]
      operationId: createVacancy
      summary: Create a vacancy
      description: |
        `company_name` is resolved to a company (created if it doesn't exist)
        when `company_id` is not given.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VacancyRequest"
      responses:
        "201":
          description: The created vacancy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vacancy"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: "`company_id` is not one of this user's companies (`error` is `not_found`)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
  /vacancies/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [vacancies]
      operationId: getVacancy
      summary: Get a vacancy
      responses:
        "200":
          description: The vacancy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vacancy"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [vacancies]
      operationId: updateVacancy
      summary: Replace a vacancy's fields
      description: |
        Despite the method, this is a full replacement: every field is written,
        and a field left out is cleared. `company_name` is ignored here.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VacancyRequest"
      responses:
        "200":
          description: The updated vacancy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vacancy"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
    delete:
      tags: [vacancies]
      operationId: deleteVacancy
      summary: Soft-delete a vacancy
      responses:
        "204":
          description: Deleted.
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /vacancies/{id}/restore:
    parameters:
      - $ref: "#/components/parameters/ID"
    post:
      tags: [vacancies]
      operationId: restoreVacancy
      summary: Undo a soft delete
      responses:
        "204":
          description: Restored.
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ------------------------------------------------------------------ applications
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: The `token` returned by the login endpoints.
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: HttpOnly cookie set by the login endpoints. Browser clients should rely on this.
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
    Desc:
      name: desc
      in: query
      schema: { type: boolean }
    Limit:
      name: limit
      in: query
      description: Values outside 1..200 fall back to 50.
      schema: { type: integer, format: int32 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, format: int32 }
  responses:
...
    ValidationFailed:
      description: The request is malformed or fails validation (`error` is `validation_failed`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: No valid session, or wrong credentials (`error` is `unauthorized`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No such resource for this user (`error` is `not_found`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The request conflicts with current state (`error` is `conflict`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnsupportedMediaType:
      description: The body was not sent as `application/json` (`error` is `unsupported_media_type`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error, message]
      properties:
        error:
          type: string
          enum:
            - validation_failed
            - unauthorized
            - unsupported_media_type
            - forbidden
            - not_found
            - conflict
            - internal_error
        message:
          type: string

    # ---------------------------------------------------------------- auth
...
    WorkMode:
      type: string
      enum: [onsite, hybrid, remote]
    EmploymentType:
      type: string
      enum: [full_time, part_time, contract, internship]
    SalaryPeriod:
      type: string
      enum: [hour, day, month, year]
    VacancyRequest:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        company_id: { type: integer, format: int64, nullable: true }
        company_name:
          type: string
          nullable: true
          description: Create only. Used when `company_id` is not given.
        title: { type: string, minLength: 1 }
        url: { type: string, nullable: true }
        description: { type: string, nullable: true }
        location: { type: string, nullable: true }
        work_mode:
          type: string
          allOf: [{ $ref: "#/components/schemas/WorkMode" }]
          nullable: true
        employment_type:
          type: string
          allOf: [{ $ref: "#/components/schemas/EmploymentType" }]
          nullable: true
        language: { type: string, nullable: true }
        salary_min: { type: integer, format: int32, nullable: true }
        salary_max: { type: integer, format: int32, nullable: true }
        salary_currency: { type: string, nullable: true }
        salary_period:
          type: string
          allOf: [{ $ref: "#/components/schemas/SalaryPeriod" }]
          nullable: true
        source: { type: string, nullable: true }
        posted_at: { type: string, format: date, nullable: true }
    Vacancy:
      type: object
      additionalProperties: false
      required: [id, user_id, title, created_at, updated_at]
      properties:
        id: { type: integer, format: int64 }
        user_id: { type: integer, format: int64 }
        company_id: { type: integer, format: int64 }
        title: { type: string }
        url: { type: string }
        description: { type: string }
        location: { type: string }
        work_mode: { $ref: "#/components/schemas/WorkMode" }
        employment_type: { $ref: "#/components/schemas/EmploymentType" }
        language: { type: string }
        salary_min: { type: integer, format: int32 }
        salary_max: { type: integer, format: int32 }
        salary_currency: { type: string }
        salary_period: { $ref: "#/components/schemas/SalaryPeriod" }
        source: { type: string }
        posted_at: { type: string, format: date }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    VacancyList:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Vacancy" }
        total: { type: integer, format: int64 }

    # ---------------------------------------------------------------- applications
...